requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
pydantic-settings==2.0.3
python-dotenv==1.0.0
pytest==7.4.3
httpx==0.25.1
websockets==12.0
pydub==0.25.1
requests==2.31.0
python-multipart==0.0.9
uuid==1.30 